Update Turbulent_2D_Constrained_RAE2822.md
Maybe it is a trivial addition, but i think that it would be better to remark this point in order to avoid strange behavior from the optimizer. For example i was sure that the thickness of my airfoil was 0.158 while the output of SU2_GEO was 0.1578. So when i've changed my thickness constraint to 0.1578 the behavior of the optimizer  have changed too co0mpletely.